:bug: StatementInspector 프로필 분리
