+++ src/main/java/dev/steady/global/config/JpaConfig.java
@@ -1,24 +1,10 @@
 package dev.steady.global.config;
 
-import dev.steady.global.logging.MDCRequestHandlerCommentInterceptor;
-import lombok.RequiredArgsConstructor;
-import org.hibernate.cfg.AvailableSettings;
-import org.springframework.boot.autoconfigure.orm.jpa.HibernatePropertiesCustomizer;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @Configuration
 @EnableJpaAuditing
-@RequiredArgsConstructor
 public class JpaConfig {
-
-    private final MDCRequestHandlerCommentInterceptor mdcRequestHandlerCommentInterceptor;
-
-    @Bean
-    public HibernatePropertiesCustomizer hibernatePropertyConfig() {
-        return hibernateProperties ->
-                hibernateProperties.put(AvailableSettings.STATEMENT_INSPECTOR, mdcRequestHandlerCommentInterceptor);
-    }
 
 }
